+++ test/nullKeysStores.mjs
@@ -10,20 +10,18 @@ describe("stores property (null keys)", function() {
 			one: "A",
 			two: "B",
 		};
-		var {stores, ready} = webextStorageAdapter(null, {
-			storageArea: makeStorageArea(storageData),
-		});
+		var {stores, ready} = webextStorageAdapter(makeStorageArea(storageData), null);
 		return ready.then( () => {
 			assert.deepStrictEqual(new Set( Object.keys(stores) ), new Set( Object.keys(storageData) ));
 		} );
 	});
 	specify("any key produces a writable store", function() {
-		var {stores} = webextStorageAdapter(null);
+		var {stores} = webextStorageAdapter("sync", null);
 		var testing = stores.anyKey;
 		assert.ok(testing.subscribe && testing.set && testing.update);
 	});
 	specify("set is an error in strict mode", function() {
-		var {stores} = webextStorageAdapter(null);
+		var {stores} = webextStorageAdapter("sync", null);
 		assert.throws( () => { stores.anyKey = 1; } );
 	});
 	specify("store.set sends new values back", function(done) {
@@ -33,7 +31,7 @@ describe("stores property (null keys)", function() {
 			assert.equal(data[expected.key], expected.value);
 			done();
 		};
-		var {stores} = webextStorageAdapter(null, {storageArea});
+		var {stores} = webextStorageAdapter(storageArea, null);
 		stores[expected.key].set(expected.value);
 	});
 	specify("store.update sends new values back", function(done) {
@@ -43,14 +41,14 @@ describe("stores property (null keys)", function() {
 			assert.equal(data[expected.key], expected.value);
 			done();
 		};
-		var {stores} = webextStorageAdapter(null, {storageArea});
+		var {stores} = webextStorageAdapter(storageArea, null);
 		stores[expected.key].update( () => expected.value );
 	});
 	specify("for live store groups, deleted keys are removed", function(done) {
 		var expected = new Set(["I'd like to be"]);
 		var initialStorage = {nananana: "hey jude", [ [...expected][0] ]: "under the sea"};
 		var storageArea = makeStorageArea(initialStorage);
-		var {stores, ready} = webextStorageAdapter(null, {storageArea});
+		var {stores, ready} = webextStorageAdapter(storageArea, null);
 		ready.then( () => {
 			stores.nananana.subscribe( (value) => {
 				try {